third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/networkState_during_progress-expected.txt

This is a testharness.js-based test.
[FAIL] audioElement.networkState should be NETWORK_LOADING during progress event
  assert_equals: expected 2 but got 1
[FAIL] videoElement.networkState should be NETWORK_LOADING during progress event
  assert_equals: expected 2 but got 1
Harness: the test ran to completion.
